requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
gunicorn==21.2.0
python-dotenv==0.21.0